added makefile.manual for manual running make
 - to compile the project run make -f makefile.manual
 - regular Makefile is a CodeLite product and does not build everything

+++ AMS-DEMO/mpiWrapper.h
@@ -116,7 +116,7 @@ namespace Mpi {
         
         // for debugging only
         int toInt() const {
-            return (int)comm;
+            return (intptr_t)comm;
         }
 	};
